+++ src/modules/feed/components/GuidelineInput/GuidelineInput.tsx
@@ -1,11 +1,12 @@
 import React, { ChangeEvent, Component } from 'react';
-import { Form, Label, Segment } from 'semantic-ui-react';
+import { Form, Segment } from 'semantic-ui-react';
 import { Query } from '@apollo/react-components';
 import gql from 'graphql-tag';
+import enhanceWithClickOutside from 'react-click-outside';
 import { Storage } from '../../../../support/storage';
 import settings from '../../../../config/settings';
 
-import s from '../../FeedPage.module.scss';
+import s from './GuidelineInput.module.scss';
 
 const KUDO_GUIDELINE_RANGE = 5;
 
@@ -23,12 +24,14 @@ export const GET_GUIDELINES = gql`
 
 export interface GetGuideLinesResult {
   teamById: {
-    guidelines: {
-      id: string;
-      kudos: number;
-      name: string;
-    }[];
+    guidelines: Guideline[];
   };
+}
+
+export interface Guideline {
+  id: string;
+  kudos: number;
+  name: string;
 }
 
 export interface Props {
@@ -38,7 +41,7 @@ export interface Props {
 
 export interface State {
   amount: string;
-  inputFocus: boolean;
+  showGuidelines: boolean;
 }
 
 class GuidelineInput extends Component<Props, State> {
@@ -52,7 +55,7 @@ class GuidelineInput extends Component<Props, State> {
 
     this.state = {
       amount: '',
-      inputFocus: false,
+      showGuidelines: false,
     };
     this.initialState = this.state;
 
@@ -61,6 +64,13 @@ class GuidelineInput extends Component<Props, State> {
     this.selectGuideline = this.selectGuideline.bind(this);
     this.handleChange = this.handleChange.bind(this);
     this.resetState = this.resetState.bind(this);
+    this.showGuidelines = this.showGuidelines.bind(this);
+  }
+
+  handleClickOutside() {
+    this.setState({
+      showGuidelines: false,
+    });
   }
 
   handleChange(e: ChangeEvent, { value }: any) {
@@ -68,17 +78,28 @@ class GuidelineInput extends Component<Props, State> {
     this.props.handleChange(value);
   }
 
+  showGuidelines(e: Event) {
+    e.preventDefault();
+    this.setState((prevState) => ({ showGuidelines: !prevState.showGuidelines }));
+  }
+
   focusKudoInput() {
-    this.setState({ inputFocus: true });
+    this.setState({
+      showGuidelines: true,
+    });
   }
 
   blurKudoInput() {
-    this.setState({ inputFocus: false });
+    this.timeout = setTimeout(() => {
+      this.setState({
+        showGuidelines: false,
+      });
+    }, 500);
   }
 
   selectGuideline(amount: string) {
     clearTimeout(this.timeout);
-    this.setState({ amount, inputFocus: false });
+    this.setState({ amount, showGuidelines: false });
     this.props.handleChange(amount);
   }
 
@@ -88,31 +109,27 @@ class GuidelineInput extends Component<Props, State> {
 
   render() {
     return (
-      <div>
-        <Form.Field>
-          <label htmlFor="input-kudos">
-            Kudos Amount
-            <Form.Input
-              data-testid="amount-input"
-              id="input-kudos"
-              error={this.props.amountError}
-              onChange={this.handleChange}
-              onFocus={this.focusKudoInput}
-              onBlur={this.blurKudoInput}
-              placeholder="Amount of kudos"
-              name="amount"
-              type="number"
-              min="0"
-              value={this.state.amount}
-              labelPosition="left"
-            >
-              <Label basic>₭</Label>
-              <input />
-            </Form.Input>
-          </label>
+      <div id="kudos-input-container" className={s.test}>
+        <Form.Field className={s.field}>
+          <Form.Input
+            className={(this.state.showGuidelines) ? s.guideline_input_active : ''}
+            data-testid="amount-input"
+            id="input-kudos"
+            error={this.props.amountError}
+            onChange={this.handleChange}
+            onFocus={this.focusKudoInput}
+            onBlur={this.blurKudoInput}
+            placeholder="Amount of kudos"
+            name="amount"
+            type="number"
+            min="0"
+            value={this.state.amount}
+            label="Kudos Amount"
+            action={{ icon: 'info', onClick: this.showGuidelines, 'data-testid': 'guidelines-button' }}
+          />
         </Form.Field>
 
-        {this.state.inputFocus && this.state.amount && (
+        {(this.state.showGuidelines) && (
         <Query<GetGuideLinesResult>
           query={GET_GUIDELINES}
           variables={{
@@ -139,13 +156,33 @@ class GuidelineInput extends Component<Props, State> {
               );
             }
 
-            // eslint-disable-next-line max-len
-            const filteredGuidlines = data.teamById.guidelines.filter((guideline) => guideline.kudos - KUDO_GUIDELINE_RANGE < (this.state.amount || 0)
-                        && guideline.kudos + KUDO_GUIDELINE_RANGE > (this.state.amount || 0));
+            let guidelines: Guideline[];
+
+            if (this.state.amount) {
+              // eslint-disable-next-line max-len
+              guidelines = data.teamById.guidelines.filter((guideline) => guideline.kudos - KUDO_GUIDELINE_RANGE < (this.state.amount || 0)
+                                    && guideline.kudos + KUDO_GUIDELINE_RANGE > (this.state.amount || 0));
+            } else {
+              guidelines = data.teamById.guidelines;
+            }
+
+            if (guidelines.length === 0) {
+              return (
+                <Segment.Group
+                  size="tiny"
+                  className={s.guidelines}
+                >
+                  <Segment key={1}>No guidelines available</Segment>
+                </Segment.Group>
+              );
+            }
 
             return (
-              <Segment.Group size="tiny" className={s.guidelines}>
-                {filteredGuidlines.map((guideline) => (
+              <Segment.Group
+                size="tiny"
+                className={s.guidelines}
+              >
+                {guidelines.map((guideline) => (
                   <Segment
                     data-testid="guideline-row"
                     key={guideline.id}
@@ -164,4 +201,4 @@ class GuidelineInput extends Component<Props, State> {
   }
 }
 
-export default GuidelineInput;
+export default enhanceWithClickOutside(GuidelineInput);